half to bf16

+++ datatype.h
@@ -2,9 +2,9 @@
 #define DATATYPEH
 
 #include <cmath>
-#include "cuda_fp16.h"
+#include "cuda_bf16.h"
 
-using DataType = __half;
+using DataType = __nv_bfloat16;
 
 __host__ __device__ inline DataType float2datatype(float in) {
     return in;
